# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axon API",
    "description": "Read API for a personal Matrix state layer. Account-scoped resources nest under /v1/accounts/{account_id}; /v1/rooms is the cross-account aggregate.",
    "license": {
      "name": "LicenseRef-axon",
      "identifier": "LicenseRef-axon"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/accounts/{account_id}/rooms/{room_id}/events/{event_id}": {
      "put": {
        "tags": [
          "messages"
        ],
        "summary": "Edit an existing message (sends an `m.replace`).",
        "operationId": "edit_message",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "description": "Event id being edited",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Edit sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SendResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "messages"
        ],
        "summary": "Redact an event, optionally with a reason (`?reason=`).",
        "operationId": "redact_event",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "description": "Event id being redacted",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "description": "Optional human-readable reason recorded on the redaction.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Redaction sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SendResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}/rooms/{room_id}/events/{event_id}/reactions": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "React to an event with an emoji/short key.",
        "operationId": "react",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "description": "Event id being reacted to",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reaction sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SendResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/{account_id}/rooms/{room_id}/send": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Send a plain-text message to a room.",
        "operationId": "send_message",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SendResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}/rooms/{room_id}/send-media": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Send a staged media upload to a room.",
        "operationId": "send_media",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Media sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SendResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account, room, or upload not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_SendResultDto": {
        "type": "object",
        "description": "Success envelope: a 2xx body is always `{ \"data\": <T> }`.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "Result of a successful mutation: the id of the event the homeserver created\n(the message, the replacement, the redaction, or the reaction).",
            "required": [
              "event_id"
            ],
            "properties": {
              "event_id": {
                "type": "string",
                "description": "The created Matrix event id."
              }
            }
          }
        }
      },
...
      "EditRequest": {
        "type": "object",
        "description": "Request body for editing a message (`PUT …/events/{event_id}`). Replaces the\ntarget event's text via an `m.replace` relation.\n\nLike [`SendMessageRequest`], `format` + `formatted_body` are optional and must\nbe supplied together to set rich text on the replacement.",
        "required": [
          "body"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The new plain-text message body (and formatting fallback)."
          },
          "format": {
            "type": [
              "string",
              "null"
            ],
            "description": "Markup name for `formatted_body` — only `org.matrix.custom.html`. Must be\npaired with `formatted_body`."
          },
          "formatted_body": {
            "type": [
              "string",
              "null"
            ],
            "description": "The rendered HTML body. Must be paired with `format`."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The error body shape: a machine-readable `code` and a human-readable\n`message`, nested under `error`.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable code, e.g. `not_found`, `bad_request`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Safe to surface to a developer; never carries\ninternal error detail for `internal` errors."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "The error envelope: `{ \"error\": { \"code\", \"message\" } }`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "The error detail."
          }
        }
      },
...
      "ReactRequest": {
        "type": "object",
        "description": "Request body for reacting to an event (`POST …/events/{event_id}/reactions`).",
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string",
            "description": "The reaction key — typically an emoji."
          }
        }
      },
...
      "SendMediaRequest": {
        "type": "object",
        "description": "Request body for sending a staged media upload into a room\n(`POST …/rooms/{room_id}/send-media`). The `upload_id` must refer to an\nunsent staged upload owned by the path account.",
        "required": [
          "upload_id"
        ],
        "properties": {
          "caption": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional media caption. When absent, Axon uses the staged filename as the\nMatrix event body."
          },
          "reply_to": {
            "type": [
              "string",
              "null"
            ],
            "description": "Send as a reply to this event id (`m.in_reply_to`). Optional."
          },
          "thread_root": {
            "type": [
              "string",
              "null"
            ],
            "description": "Send into this thread, identified by its root event id\n(`rel_type: m.thread`). Optional; when set without `reply_to`, the\nthread root is used as the reply fallback."
          },
          "upload_id": {
            "type": "string",
            "format": "uuid",
            "description": "Server-issued staged upload id returned by `POST …/media/uploads`."
          }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "description": "Request body for sending a message (`POST …/rooms/{room_id}/send`). Sent as an\n`m.room.message`; `account_id`/`room_id` come from the path.\n\n`body` is the plain-text content (and the fallback for clients that don't\nrender formatting). To send rich text, supply `format` *and* `formatted_body`\ntogether: `format` must be `org.matrix.custom.html` and `formatted_body` the\nrendered HTML. Axon carries them verbatim — it never interprets `body` as\nMarkdown.",
        "required": [
          "body"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The plain-text message body (and formatting fallback)."
          },
          "format": {
            "type": [
              "string",
              "null"
            ],
            "description": "Markup name for `formatted_body` — only `org.matrix.custom.html`. Must be\npaired with `formatted_body`."
          },
          "formatted_body": {
            "type": [
              "string",
              "null"
            ],
            "description": "The rendered HTML body. Must be paired with `format`."
          },
          "reply_to": {
            "type": [
              "string",
              "null"
            ],
            "description": "Send as a reply to this event id (`m.in_reply_to`). Optional."
          },
          "thread_root": {
            "type": [
              "string",
              "null"
            ],
            "description": "Send into this thread, identified by its root event id\n(`rel_type: m.thread`). Optional; when set with `reply_to`, the reply is\nscoped to the thread. Without `reply_to`, this is a thread member, not a\nreply."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    }
  ],
...
}